weather info is being outputed in farm details recycler

+++ app/src/main/java/com/example/beekeeper/presenter/model/weather/WeatherInfoUI.kt
@@ -1,17 +1,15 @@
 package com.example.beekeeper.presenter.model.weather
 
-import com.example.beekeeper.domain.model.weather.MainInfo
-
 data class WeatherInfoUI(
     val main: MainInfoUI,
     val weather: List<WeatherUI>
 )
 
 data class MainInfoUI(
-    val temp: Double,
-    val feelsLike: Double,
-    val tempMin: Double,
-    val tempMax: Double,
+    val temp: Int,
+    val feelsLike: Int,
+    val tempMin: Int,
+    val tempMax: Int,
     val pressure: Int,
     val humidity: Int,
     val seaLevel: Int? = null, // Optional
